JSONRPC.py added.
This new class is the lowest level so far. It's the responsible for handling
HTTP requests and return the JSON response.

It can ask for all or specifics permissions, allowing a fine control of XBMC
remote calls.
A ping method checks for connectivity. And version method return the XMBC
system's version.

The documentation is coming, I promise. For now, the code is very well written.