It's alive!
Yesterday I added another repository, called dashboard. The purpose is to
initiate the website that will control all the clients

This commit is the back-end adaptation, which includes refactoring in the model
file. Now it imports Node class and gives real-time data from the clients.

So awesome! I want to high-five myself :)